+++ src/main/java/unicon/matthews/security/model/token/RefreshToken.java
@@ -8,7 +8,7 @@ import java.util.Optional;
 
 import org.springframework.security.authentication.BadCredentialsException;
 
-import unicon.matthews.security.exceptions.JwtExpiredTokenException;
+import unicon.matthews.security.exception.JwtExpiredTokenException;
 import unicon.matthews.security.model.Scopes;
 
 /**
